+++ packages/utils/src/script.ts
@@ -7,7 +7,6 @@ import {
   encodeAuthenticationInstructions,
   decodeAuthenticationInstructions,
 } from '@bitauth/libauth';
-import { decodeInt, encodeInt } from './data.js';
 import OptimisationsEquivFile from './cashproof-optimisations.js';
 
 export const Op = OpcodesBch2023;
@@ -142,49 +141,8 @@ function getPushDataOpcode(data: Uint8Array): Uint8Array {
   throw Error('Pushdata too large');
 }
 
-/**
- * When cutting out the tx.bytecode preimage variable, the compiler does not know
- * the size of the final redeem scrip yet, because the constructor parameters still
- * need to get added. Because of this it does not know whether the VarInt is 1 or 3
- * bytes. During compilation, an OP_NOP is added at the spot where the bytecode is
- * cut out. This function replaces that OP_NOP and adds either 1 or 3 to the cut to
- * additionally cut off the VarInt.
- *
- * @param script incomplete redeem script
- * @returns completed redeem script
- */
-export function replaceBytecodeNop(script: Script): Script {
-  const index = script.findIndex((op) => op === Op.OP_NOP);
-  if (index < 0) return script;
-
-  // Remove the OP_NOP
-  script.splice(index, 1);
-
-  // Retrieve size of current OP_SPLIT
-  let oldCut = script[index];
-  if (oldCut instanceof Uint8Array) {
-    oldCut = Number(decodeInt(oldCut));
-  } else if (oldCut === Op.OP_0) {
-    oldCut = 0;
-  } else if (oldCut >= Op.OP_1 && oldCut <= Op.OP_16) {
-    oldCut -= 80;
-  } else {
-    return script;
-  }
-
-  // Update the old OP_SPLIT by adding either 1 or 3 to it
-  script[index] = encodeInt(BigInt(oldCut + 1));
-  const bytecodeSize = calculateBytesize(script);
-  if (bytecodeSize > 252) {
-    script[index] = encodeInt(BigInt(oldCut + 3));
-  }
-
-  // Minimally encode
-  return asmToScript(scriptToAsm(script));
-}
-
 export function generateRedeemScript(baseScript: Script, encodedConstructorArgs: Script): Script {
-  return replaceBytecodeNop([...encodedConstructorArgs.slice().reverse(), ...baseScript]);
+  return [...encodedConstructorArgs.slice().reverse(), ...baseScript];
 }
 
 export function optimiseBytecode(script: Script, runs: number = 1000): Script {